+++ CHANGELOG.md
@@ -11,5 +11,6 @@ and this project adheres to [Haskell Package Versioning Policy](https://pvp.hask
 
 - First version. Released on an unsuspecting world.
 - Let users extend the built-in diff types with custom diffs via the `SpecialDiff` class in [#9](https://github.com/fpringle/generic-diff/pull/9).
+- Add example implementations of `SpecialDiff` for `containers` types in [#10](https://github.com/fpringle/generic-diff/pull/10).
 
 [unreleased]: https://github.com/fpringle/generic-diff/compare/74b5028...HEAD